Se creo un simulador de gestor de ventanas, en un formulario principal se convierte en un contenedor para abrir ventanas de forma simultanea
