+++ Practica_1_CMD/Practica_1_CMD.csproj
@@ -55,11 +55,17 @@
     <Compile Include="Bateria.Designer.cs">
       <DependentUpon>Bateria.cs</DependentUpon>
     </Compile>
-    <Compile Include="cmd.cs">
+    <Compile Include="Cmd.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="cmd.Designer.cs">
-      <DependentUpon>cmd.cs</DependentUpon>
+    <Compile Include="Cmd.Designer.cs">
+      <DependentUpon>Cmd.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Inicio.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Inicio.Designer.cs">
+      <DependentUpon>Inicio.cs</DependentUpon>
     </Compile>
     <Compile Include="Descargar.cs">
       <SubType>Form</SubType>
@@ -97,7 +103,10 @@
       <DependentUpon>Bateria.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="cmd.resx">
-      <DependentUpon>cmd.cs</DependentUpon>
+      <DependentUpon>Cmd.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Inicio.resx">
+      <DependentUpon>Inicio.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Descargar.resx">
       <DependentUpon>Descargar.cs</DependentUpon>
